refactor: extract prescription PDF composer
